MyUnlock.exe & payload.exe
Implemented MyUnlock. Still needs to be changed to a GET request from the server for the privateKey.

Changed gathering of all filenames to also include traversal of subdirectories.

Fixed bug where payload was encrypting itself. Issue was that filepath comes in the form of ".payload.exe"

Created .exe files for both payload and MyUnlock